rm struct
