+++ codex-rs/core/src/tools/mod.rs
@@ -1,3 +1,4 @@
+mod approvals;
 pub(crate) mod code_mode;
 pub(crate) mod context;
 pub(crate) mod events;
